+++ plugins/woocommerce-admin/client/task-list/task-item.tsx
@@ -4,8 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { Icon, check } from '@wordpress/icons';
 import { Button, Tooltip } from '@wordpress/components';
-import { Text } from '@woocommerce/experimental';
-import { __experimentalListItem as ListItem } from '@woocommerce/components';
+import { Text, ListItem } from '@woocommerce/experimental';
 import NoticeOutline from 'gridicons/dist/notice-outline';
 import classnames from 'classnames';