+++ plugins/woocommerce-admin/packages/components/src/list/index.js
@@ -112,7 +112,3 @@ List.propTypes = {
 };
 
 export default List;
-
-export { ExperimentalListItem } from './experimental-list-item';
-export { ExperimentalList } from './experimental-list';
-export { ExperimentalCollapsibleList } from './collapsible-list';