+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -3,9 +3,8 @@
 -   SelectControl: automatically scroll to selected options when list is displayed. #6906
 -   SelectControl: no longer auto selects on rendering list. #6906
 -   Make `Search` accept synchronous `autocompleter.options`. #6884
--   Add new (experimental) collapsible list item to collapse list items. #6869
 -   SelectControl: fix display of multiple selections without inline tags. #6862
--   Add new (experimental) list, and add depreciation notice for the current list. #6787
+-   Add depreciation notice for the current list. #6787
 -   Force `<SearchListItem>` form elements id to be unique. #6871
 -   Add `controlId` and `name` props to `<SearchListItem>`. #6871
 -   Minor styling tweaks and fixes to `<SearchListcontrol>`. #6871