+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
--   Export component ExperimentalList and ExperimentalListItem as List and ListItem.
+-   Add new (experimental) collapsible list item to collapse list items. #6869
+-   Add new (experimental) list. #6787
 
 # 1.0.0